+++ .github/HISTORY.md
@@ -1,4 +1,8 @@
 # React-Images-With-Youtube
 
+### 0.1.1 / 2017-04-28
+- Add wrapper around YouTube frame to behave responsively for small screen widths
+- Add example using a `youtubeVideoId` in the Lightbox.
+
 ### 0.1.0 / 2017-04-19
 - Initial release of fork with Youtube Video support